[fat] Retain the pageframes of the RAMDISK
Generally, all the pageframes (pages in the physical memory) must be retained,
otherwise pdir_destroy() will call kfree2() if it finds them mapped in a given
pdir and with ref-count=0. Of course, that would cause a panic, because the
memory has never been allocated with kmalloc() in the first place.

Therefore, we have to pin these pageframes by keeping their ref-count >= 1.
Until now, it was not strictly necessary to do so for the ramdisk but, because
of some upcoming changes, that will becomre required.